third_party/WebKit/LayoutTests/svg/animations/currentTime-use-counter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg id="svgRoot" width="400" height="400">
</svg>
<script>
'use strict';

// From UseCounter.h
var SVGSMILPausing = 1506;
var SVGSMILCurrentTime = 1507;

test(() => {
    var svgRoot = document.getElementById('svgRoot');

    assert_false(internals.isUseCounted(document, SVGSMILPausing));
    svgRoot.pauseAnimations();
    assert_true(internals.isUseCounted(document, SVGSMILPausing));

    assert_false(internals.isUseCounted(document, SVGSMILCurrentTime));
    svgRoot.setCurrentTime(6000);
    assert_true(internals.isUseCounted(document, SVGSMILCurrentTime));
}, 'setCurrentTime is use counted');

</script>
