third_party/WebKit/LayoutTests/platform/win/svg/animations/currentTime-use-counter-expected.txt

This is a testharness.js-based test.
PASS setCurrentTime is use counted 
Harness: the test ran to completion.
